Client now should now correctly allow registering, logging in, starting the Game UI as the color the server instructs, and moving the foe's pieces as the server instructs.  The EchoServer class from simple-chat was used to ensure that GameClient is actually sending out messages correctly.
The client also has a login UI that prompts the user if there is no server at their address and port number.

The Client and gameUI now store who is the player and enemy using the friend and foe variables.  The GUI class constructor was changed from GUI() to GUI(Spacestate friend, SpaceState foe) to facilitate this change.

TODO: Allow for mouse moves on the GUI.